$ Fixed a handful of errors left from using Notepad.
