+++ src/Java/gtPlusPlus/core/item/general/chassis/itemBoilerChassis.java
@@ -1,6 +1,19 @@
+package gtPlusPlus.core.item.general.chassis;
+import java.util.List;
+
+import cpw.mods.fml.common.registry.GameRegistry;
+import gtPlusPlus.core.creative.AddToCreativeTab;
+import gtPlusPlus.core.lib.CORE;
+import gtPlusPlus.core.util.Utils;
+import net.minecraft.client.renderer.texture.IIconRegister;
+import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.IIcon;
+
 public class itemBoilerChassis extends Item {
 
-	public IIcon[] icon = new IIcon[1];
+	public IIcon[] icons = new IIcon[1];
 
 	public itemBoilerChassis() {
 		super();
@@ -54,16 +67,16 @@ public class itemBoilerChassis extends Item {
 	public int getColorFromItemStack(final ItemStack stack, int HEX_OxFFFFFF) {
 		int meta = stack.getItemDamage();
 		if (meta == 0){
-			HEX_OxFFFFFF = Utils.rgbToHex(10,110,30);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(10,110,30);
 		}
 		else if (meta == 1){
-			HEX_OxFFFFFF = Utils.rgbToHex(150,180,35);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(150,180,35);
 		}	
 		else if (meta == 2){
-			HEX_OxFFFFFF = Utils.rgbToHex(200,85,40);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(200,85,40);
 		}
 		else if (meta == 3){
-			HEX_OxFFFFFF = Utils.rgbToHex(255,100,50);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(255,100,50);
 		}
 		return HEX_OxFFFFFF;
 	}
