+++ src/Java/gtPlusPlus/core/item/general/chassis/itemDehydratorCoil.java
@@ -1,13 +1,25 @@
-package com.bedrockminer.tutorial.item;
+package gtPlusPlus.core.item.general.chassis;
 
-public class itemDehydratorCoilWire extends Item {
+import java.util.List;
 
-	public IIcon[] icon = new IIcon[1];
+import cpw.mods.fml.common.registry.GameRegistry;
+import gtPlusPlus.core.creative.AddToCreativeTab;
+import gtPlusPlus.core.lib.CORE;
+import gtPlusPlus.core.util.Utils;
+import net.minecraft.client.renderer.texture.IIconRegister;
+import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.IIcon;
 
-	public itemDehydratorCoilWire() {
+public class itemDehydratorCoil extends Item {
+
+	public IIcon[] icons = new IIcon[1];
+
+	public itemDehydratorCoil() {
 		super();
 		this.setHasSubtypes(true);
-		String unlocalizedName = "itemDehydratorCoilWire";
+		String unlocalizedName = "itemDehydratorCoil";
 		this.setUnlocalizedName(unlocalizedName);
 		this.setCreativeTab(AddToCreativeTab.tabMisc);
 		GameRegistry.registerItem(this, unlocalizedName);
@@ -15,7 +27,7 @@ public class itemDehydratorCoilWire extends Item {
 
 	@Override
 	public void registerIcons(IIconRegister reg) {
-		this.icons[0] = reg.registerIcon(CORE.MODID + ":" + "itemDehydratorCoilWire");
+		this.icons[0] = reg.registerIcon(CORE.MODID + ":" + "itemDehydratorCoil");
 	}
 
 	@Override
@@ -37,7 +49,7 @@ public class itemDehydratorCoilWire extends Item {
 
 	@Override
 	public String getItemStackDisplayName(final ItemStack tItem) {
-		String itemName = "Coil Wire";
+		String itemName = "Dehydrator Coil";
 		String suffixName = "";
 		if (tItem.getItemDamage() == 0){
 			suffixName = " [EV]";
@@ -59,16 +71,16 @@ public class itemDehydratorCoilWire extends Item {
 	public int getColorFromItemStack(final ItemStack stack, int HEX_OxFFFFFF) {
 		int meta = stack.getItemDamage();
 		if (meta == 0){
-			HEX_OxFFFFFF = Utils.rgbToHex(10,110,30);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(10,110,30);
 		}
 		else if (meta == 1){
-			HEX_OxFFFFFF = Utils.rgbToHex(150,180,35);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(150,180,35);
 		}	
 		else if (meta == 2){
-			HEX_OxFFFFFF = Utils.rgbToHex(200,85,40);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(200,85,40);
 		}
 		else if (meta == 3){
-			HEX_OxFFFFFF = Utils.rgbToHex(255,100,50);
+			HEX_OxFFFFFF = Utils.rgbtoHexValue(255,100,50);
 		}
 		return HEX_OxFFFFFF;
 	}
